Refactor command scrpits
