+++ bin/commands/init.js
