+++ bin/commands/openapi-fetch.js
